+++ example/personorm/create.go
@@ -3,5 +3,5 @@ package personorm
 
 func (c *TCreate) String() string {
 	// Create statement has a line for each variable with it's name and it's type.
-	return `CREATE TABLE 'person' ( 'name' VARCHAR(255), 'age' BIGINT )`
+	return `CREATE TABLE 'person' ( 'name' VARCHAR(255), 'age' INT )`
 }
